+++ src/command_handler.rs
@@ -1,3 +1,3 @@
 use crate::{Error, Intent, Result};
 
-pub type CommandHandler<C> = fn(err: Error, intent: &Intent, context: &mut C) -> Result<usize>;
+pub type CommandHandler<C> = fn(err: Error, intent: &Intent, context: &mut C) -> Result<i32>;
